# openapi/openapi.bundle.yaml
openapi: 3.1.0
info:
  title: moly-backend API
  version: 1.0.0
  summary: Moly 앱 백엔드 API 계약
  description: '이 문서는 moly-backend 공개 HTTP API의 확정 계약이다. 호환성에 영향을 주는 변경은 Semantic Versioning으로 관리한다. 보호
    API는 Supabase access token Bearer 인증을 사용한다. `/openapi.json`과 `/docs`는 local·development에서만 열리고 staging·production에서는
    등록되지 않는다. `/redoc`은 항상 비활성화되어 있다.


    애플리케이션 오류, 요청 검증 오류, 등록되지 않은 path의 404, 등록된 path의 미지원 method 405, 미처리 예외 500은 모두 `ErrorEnvelope`로
    반환된다. 깨진 JSON과 비객체 body를 포함한 요청 검증 실패는 400이 아니라 422다.


    날짜 경계는 도메인마다 다르다. 채팅 token과 일기 귀속의 `activity_date`는 사용자 local 04:00, 출석·루틴·광고 보상의 기준일은 사용자 local 00:00이다.
    `/charging-station.activity_date` 필드명은 레거시이며 실제 의미는 00:00 경계 reward date다.


    local·development 환경의 명시 플래그로만 등록되는 `/dev/*` 라우터와 moly-auth 소유 계정 API(`/me` 포함)는 이 계약 범위 밖이다.'
paths:
...
  /routines:
    get:
      tags:
      - routine
      summary: 루틴 목록 조회
      operationId: listRoutines
      responses:
        '200':
          description: 루틴 목록
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoutineListResponse'
          headers:
            X-App-Local-Date:
              description: 실제 계산에 사용한 현지 날짜
              schema:
                type: string
                format: date
            X-App-Served-At:
              description: 단일 요청 서버 UTC 시각
              schema:
                type: string
                format: date-time
            X-App-Day-Ends-At:
              description: 다음 현지 달력일 자정의 UTC 시각
              schema:
                type: string
                format: date-time
        '500':
          $ref: '#/components/responses/Internal'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/ProfileNotFound'
        '422':
          $ref: '#/components/responses/Validation'
      security:
      - SupabaseBearer: []
      description: soft-delete되지 않은 본인 routine을 created_at, id 오름차순으로 모두 반환한다. pagination은 없다. completed_today·skipped_today는
        요청의 현지 날짜 기준이다.
      parameters:
      - name: X-App-Timezone
        in: header
        required: false
        description: 현재 기기의 IANA 시간대. 생략 시 저장 시간대. 잘못된 값은 APP_TIMEZONE_INVALID.
        schema:
          type: string
          minLength: 1
          maxLength: 64
    post:
      tags:
      - routine
      summary: 루틴 생성
      operationId: createRoutine
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoutineRequest'
        required: true
      responses:
        '201':
          description: 생성된 루틴
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoutineResponse'
          headers:
            X-App-Local-Date:
              description: 실제 계산에 사용한 현지 날짜
              schema:
                type: string
                format: date
            X-App-Served-At:
              description: 단일 요청 서버 UTC 시각
              schema:
                type: string
                format: date-time
            X-App-Day-Ends-At:
              description: 다음 현지 달력일 자정의 UTC 시각
              schema:
                type: string
                format: date-time
        '500':
          $ref: '#/components/responses/Internal'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/ProfileNotFound'
        '422':
          $ref: '#/components/responses/Validation'
      security:
      - SupabaseBearer: []
      description: days_of_week는 ISO 월=1…일=7 중복 없는 1개 이상이며 저장 전에 정렬된다. frequency_per_week는 배열 길이로 파생된다.
        icon 기본값은 seedling, color 기본값은 peach다. 요일 이력은 요청의 현지 날짜부터 시작한다.
      parameters:
      - name: X-App-Timezone
        in: header
        required: false
        description: 현재 기기의 IANA 시간대. 생략 시 저장 시간대. 잘못된 값은 APP_TIMEZONE_INVALID.
        schema:
          type: string
          minLength: 1
          maxLength: 64
  /routines/history:
    get:
      tags:
      - routine
      summary: 날짜별 루틴 기록 조회
      operationId: getRoutineHistory
      responses:
        '200':
          description: 선택일 루틴 기록
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoutineHistoryResponse'
          headers:
            X-App-Local-Date:
              description: 실제 계산에 사용한 현지 날짜
              schema:
                type: string
                format: date
            X-App-Served-At:
              description: 단일 요청 서버 UTC 시각
              schema:
                type: string
                format: date-time
            X-App-Day-Ends-At:
              description: 다음 현지 달력일 자정의 UTC 시각
              schema:
                type: string
                format: date-time
        '500':
          $ref: '#/components/responses/Internal'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/ProfileNotFound'
        '422':
          $ref: '#/components/responses/Validation'
      security:
      - SupabaseBearer: []
      description: 선택일에 예정이었던 루틴(스킵 포함, 나중에 삭제된 루틴 포함)과 선택일에 완료한 루틴을 반환한다. 예정 여부와 days_of_week는 선택일에 적용되던
        요일 이력으로 정한다. 미래 날짜는 422. 날짜 헤더는 선택 날짜가 아니라 요청 시점의 실제 오늘이다. 읽기 전용이며 조회 기간에 30일 제한이 없다.
      parameters:
      - name: date
        in: query
        required: true
        schema:
          type: string
          format: date
        description: YYYY-MM-DD 달력 날짜. 실제 오늘 이하.
      - name: X-App-Timezone
        in: header
        required: false
        description: 현재 기기의 IANA 시간대. 생략 시 저장 시간대. 잘못된 값은 APP_TIMEZONE_INVALID.
        schema:
          type: string
          minLength: 1
          maxLength: 64
  /routines/{routine_id}:
    patch:
      tags:
      - routine
      summary: 루틴 수정
      operationId: updateRoutine
      security:
      - SupabaseBearer: []
      parameters:
      - name: routine_id
        in: path
        required: true
        schema:
          type: string
          title: Routine Id
        description: UUID string. 형식이 잘못돼도 404로 처리한다.
      - name: X-App-Timezone
        in: header
        required: false
        description: 현재 기기의 IANA 시간대. 생략 시 저장 시간대. 잘못된 값은 APP_TIMEZONE_INVALID.
        schema:
          type: string
          minLength: 1
          maxLength: 64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchRoutineRequest'
      responses:
        '204':
          description: 수정 완료; body 없음
        '500':
          $ref: '#/components/responses/Internal'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/Validation'
        '404':
          description: UUID 형식 오류, 없음, 타인 소유 또는 soft-delete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              examples:
                notFound:
                  summary: resource 없음
                  value:
                    error:
                      code: NOT_FOUND
                      message: 루틴을 찾을 수 없어요.
                      details: {}
      description: body 자체는 필수지만 `{}`는 no-op 204다. name/reminder_enabled/days_of_week/icon/color의 null은
        변경 없음. reminder_time은 생략하면 변경 없고 명시적 null이면 제거한다. days_of_week를 바꾸면 frequency_per_week도 길이로 갱신하고,
        요청의 현지 날짜부터 새 요일을 적용한다. 그 전날까지의 기록은 바뀌지 않는다.
    delete:
      tags:
      - routine
      summary: 루틴 삭제
      operationId: deleteRoutine
      security:
      - SupabaseBearer: []
      parameters:
      - name: routine_id
        in: path
        required: true
        schema:
          type: string
          title: Routine Id
        description: UUID string. 형식이 잘못돼도 404로 처리한다.
      - name: X-App-Timezone
        in: header
        required: false
        description: 현재 기기의 IANA 시간대. 생략 시 저장 시간대. 잘못된 값은 APP_TIMEZONE_INVALID.
        schema:
          type: string
          minLength: 1
          maxLength: 64
      responses:
        '204':
          description: 삭제 완료; body 없음
        '500':
          $ref: '#/components/responses/Internal'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: UUID 형식 오류, 없음, 타인 소유 또는 soft-delete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              examples:
                notFound:
                  summary: resource 없음
                  value:
                    error:
                      code: NOT_FOUND
                      message: 루틴을 찾을 수 없어요.
                      details: {}
      description: deleted_at과 요청의 현지 날짜(deleted_on)를 기록하는 soft delete다. 그날부터 예정에서 빠지고 지난 기록은 남는다.
  /routines/{routine_id}/complete:
    post:
      tags:
      - routine
      summary: 오늘 루틴 완료
      operationId: completeRoutine
      security:
      - SupabaseBearer: []
      parameters:
      - name: routine_id
        in: path
        required: true
        schema:
          type: string
          title: Routine Id
        description: UUID string. 형식이 잘못돼도 404로 처리한다.
      - name: X-App-Timezone
        in: header
        required: false
        description: 현재 기기의 IANA 시간대. 생략 시 저장 시간대. 잘못된 값은 APP_TIMEZONE_INVALID.
        schema:
          type: string
          minLength: 1
          maxLength: 64
      responses:
        '200':
          description: 완료 상태와 오늘 전체 완료 수
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoutineCompleteResponse'
          headers:
            X-App-Local-Date:
              description: 실제 계산에 사용한 현지 날짜
              schema:
                type: string
                format: date
            X-App-Served-At:
              description: 단일 요청 서버 UTC 시각
              schema:
                type: string
                format: date-time
            X-App-Day-Ends-At:
              description: 다음 현지 달력일 자정의 UTC 시각
              schema:
                type: string
                format: date-time
        '500':
          $ref: '#/components/responses/Internal'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: profile이 없거나 routine UUID/소유/삭제 상태가 유효하지 않음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              examples:
                profile:
                  summary: 프로필 없음
                  value:
                    error:
                      code: NOT_FOUND
                      message: 프로필을 찾을 수 없어요.
                      details: {}
                routine:
                  summary: 루틴 없음
                  value:
                    error:
                      code: NOT_FOUND
                      message: 루틴을 찾을 수 없어요.
                      details: {}
        '422':
          $ref: '#/components/responses/Validation'
      description: local 00:00 reward date 기준 완료를 insert하며 unique conflict를 무시하므로 멱등이다. 같은 날 스킵이 있으면 지운다.
    delete:
      tags:
      - routine
      summary: 오늘 루틴 완료 취소
      operationId: uncompleteRoutine
      security:
      - SupabaseBearer: []
      parameters:
      - name: routine_id
        in: path
        required: true
        schema:
          type: string
          title: Routine Id
        description: UUID string. 형식이 잘못돼도 404로 처리한다.
      - name: X-App-Timezone
        in: header
        required: false
        description: 현재 기기의 IANA 시간대. 생략 시 저장 시간대. 잘못된 값은 APP_TIMEZONE_INVALID.
        schema:
          type: string
          minLength: 1
          maxLength: 64
      responses:
        '204':
          description: 완료 취소; body 없음
          headers:
            X-App-Local-Date:
              description: 실제 계산에 사용한 현지 날짜
              schema:
                type: string
                format: date
            X-App-Served-At:
              description: 단일 요청 서버 UTC 시각
              schema:
                type: string
                format: date-time
            X-App-Day-Ends-At:
              description: 다음 현지 달력일 자정의 UTC 시각
              schema:
                type: string
                format: date-time
        '500':
          $ref: '#/components/responses/Internal'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: profile이 없거나 routine UUID/소유/삭제 상태가 유효하지 않음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              examples:
                profile:
                  summary: 프로필 없음
                  value:
                    error:
                      code: NOT_FOUND
                      message: 프로필을 찾을 수 없어요.
                      details: {}
                routine:
                  summary: 루틴 없음
                  value:
                    error:
                      code: NOT_FOUND
                      message: 루틴을 찾을 수 없어요.
                      details: {}
        '422':
          $ref: '#/components/responses/Validation'
      description: 해당 reward date 완료 행이 없어도 204다.
  /routines/{routine_id}/statistics:
    get:
      tags:
      - routine
      summary: 루틴 통계 조회
      operationId: getRoutineStatistics
      security:
      - SupabaseBearer: []
      parameters:
      - name: routine_id
        in: path
        required: true
        schema:
          type: string
          title: Routine Id
        description: UUID string. 형식이 잘못돼도 404로 처리한다.
      - name: X-App-Timezone
        in: header
        required: false
        description: 현재 기기의 IANA 시간대. 생략 시 저장 시간대. 잘못된 값은 APP_TIMEZONE_INVALID.
        schema:
          type: string
          minLength: 1
          maxLength: 64
      responses:
        '200':
          description: 루틴 통계
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoutineStatisticsResponse'
          headers:
            X-App-Local-Date:
              description: 실제 계산에 사용한 현지 날짜
              schema:
                type: string
                format: date
            X-App-Served-At:
              description: 단일 요청 서버 UTC 시각
              schema:
                type: string
                format: date-time
            X-App-Day-Ends-At:
              description: 다음 현지 달력일 자정의 UTC 시각
              schema:
                type: string
                format: date-time
        '500':
          $ref: '#/components/responses/Internal'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: profile이 없거나 routine UUID/소유/삭제 상태가 유효하지 않음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              examples:
                profile:
                  summary: 프로필 없음
                  value:
                    error:
                      code: NOT_FOUND
                      message: 프로필을 찾을 수 없어요.
                      details: {}
                routine:
                  summary: 루틴 없음
                  value:
                    error:
                      code: NOT_FOUND
                      message: 루틴을 찾을 수 없어요.
                      details: {}
        '422':
          $ref: '#/components/responses/Validation'
      description: streak는 지정 요일이 아니라 오늘부터 역방향 연속 calendar day 수다. completion_rate는 최근 28일 완료 수/(요일 수×4),
        상한 1.0, 소수 둘째 자리 반올림이다.
  /routines/stats:
    get:
      tags:
      - routine
      summary: 루틴 기록 통계
      operationId: getRoutineStats
      security:
      - SupabaseBearer: []
      description: 날짜는 모두 요청의 현지 날짜다. 인스턴스는 그날 완료했거나, 예정이었고 스킵하지 않은 (루틴, 날짜)다. summary는 삭제된 루틴을 포함한 전
        기간 값이고 days는 from..min(to, today)의 날짜별 인스턴스·완료 수다. from ≤ to이고 기간은 371일 이하여야 하며 to가 today 뒤면 today로
        자른다.
      parameters:
      - name: from
        in: query
        required: true
        schema:
          type: string
          format: date
        description: YYYY-MM-DD 시작 날짜(포함).
      - name: to
        in: query
        required: true
        schema:
          type: string
          format: date
        description: YYYY-MM-DD 끝 날짜(포함). from부터 371일 이하.
      - name: X-App-Timezone
        in: header
        required: false
        description: 현재 기기의 IANA 시간대. 생략 시 저장 시간대. 잘못된 값은 APP_TIMEZONE_INVALID.
        schema:
          type: string
          minLength: 1
          maxLength: 64
      responses:
        '200':
          description: 루틴 기록 통계
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoutineStatsResponse'
          headers:
            X-App-Local-Date:
              description: 실제 계산에 사용한 현지 날짜
              schema:
                type: string
                format: date
            X-App-Served-At:
              description: 단일 요청 서버 UTC 시각
              schema:
                type: string
                format: date-time
            X-App-Day-Ends-At:
              description: 다음 현지 달력일 자정의 UTC 시각
              schema:
                type: string
                format: date-time
        '500':
          $ref: '#/components/responses/Internal'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/ProfileNotFound'
        '422':
          $ref: '#/components/responses/Validation'
  /routines/{routine_id}/skip:
    post:
      tags:
      - routine
      summary: 오늘 루틴 건너뛰기
      operationId: skipRoutine
      security:
      - SupabaseBearer: []
      description: 요청의 현지 날짜에 스킵을 기록한다. 이미 스킵이면 그대로 204다. 오늘 완료한 루틴이면 409.
      parameters:
      - name: routine_id
        in: path
        required: true
        schema:
          type: string
          title: Routine Id
        description: UUID string. 형식이 잘못돼도 404로 처리한다.
      - name: X-App-Timezone
        in: header
        required: false
        description: 현재 기기의 IANA 시간대. 생략 시 저장 시간대. 잘못된 값은 APP_TIMEZONE_INVALID.
        schema:
          type: string
          minLength: 1
          maxLength: 64
      responses:
        '204':
          description: 스킵 기록; body 없음
          headers:
            X-App-Local-Date:
              description: 실제 계산에 사용한 현지 날짜
              schema:
                type: string
                format: date
            X-App-Served-At:
              description: 단일 요청 서버 UTC 시각
              schema:
                type: string
                format: date-time
            X-App-Day-Ends-At:
              description: 다음 현지 달력일 자정의 UTC 시각
              schema:
                type: string
                format: date-time
        '500':
          $ref: '#/components/responses/Internal'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: profile이 없거나 routine UUID/소유/삭제 상태가 유효하지 않음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              examples:
                profile:
                  summary: 프로필 없음
                  value:
                    error:
                      code: NOT_FOUND
                      message: 프로필을 찾을 수 없어요.
                      details: {}
                routine:
                  summary: 루틴 없음
                  value:
                    error:
                      code: NOT_FOUND
                      message: 루틴을 찾을 수 없어요.
                      details: {}
        '409':
          description: 오늘 이미 완료한 루틴
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              examples:
                completed:
                  summary: 완료한 루틴
                  value:
                    error:
                      code: ROUTINE_ALREADY_COMPLETED
                      message: 완료한 루틴은 건너뛸 수 없어요.
                      details: {}
        '422':
          $ref: '#/components/responses/Validation'
    delete:
      tags:
      - routine
      summary: 오늘 루틴 건너뛰기 취소
      operationId: unskipRoutine
      security:
      - SupabaseBearer: []
      description: 요청의 현지 날짜의 스킵을 지운다. 스킵이 없어도 204다.
      parameters:
      - name: routine_id
        in: path
        required: true
        schema:
          type: string
          title: Routine Id
        description: UUID string. 형식이 잘못돼도 404로 처리한다.
      - name: X-App-Timezone
        in: header
        required: false
        description: 현재 기기의 IANA 시간대. 생략 시 저장 시간대. 잘못된 값은 APP_TIMEZONE_INVALID.
        schema:
          type: string
          minLength: 1
          maxLength: 64
      responses:
        '204':
          description: 스킵 취소; body 없음
          headers:
            X-App-Local-Date:
              description: 실제 계산에 사용한 현지 날짜
              schema:
                type: string
                format: date
            X-App-Served-At:
              description: 단일 요청 서버 UTC 시각
              schema:
                type: string
                format: date-time
            X-App-Day-Ends-At:
              description: 다음 현지 달력일 자정의 UTC 시각
              schema:
                type: string
                format: date-time
        '500':
          $ref: '#/components/responses/Internal'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: profile이 없거나 routine UUID/소유/삭제 상태가 유효하지 않음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              examples:
                profile:
                  summary: 프로필 없음
                  value:
                    error:
                      code: NOT_FOUND
                      message: 프로필을 찾을 수 없어요.
                      details: {}
                routine:
                  summary: 루틴 없음
                  value:
                    error:
                      code: NOT_FOUND
                      message: 루틴을 찾을 수 없어요.
                      details: {}
        '422':
          $ref: '#/components/responses/Validation'
  /routine-templates:
    get:
      tags:
      - routine
      summary: 루틴 템플릿 목록
      operationId: listRoutineTemplates
      security:
      - SupabaseBearer: []
      description: 서버가 관리하는 활성 카테고리·템플릿을 sort_order 순으로 프로필 언어로 현지화해 반환한다.
      responses:
        '200':
          description: 템플릿 목록
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoutineTemplatesResponse'
        '500':
          $ref: '#/components/responses/Internal'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/ProfileNotFound'
  /routine-templates/selection:
    post:
      tags:
      - routine
      summary: 루틴 템플릿 선택
      operationId: selectRoutineTemplates
      security:
      - SupabaseBearer: []
      description: 고른 템플릿으로 루틴을 카테고리·템플릿 순서대로 만들고 이번에 만든 루틴만 반환한다. 이미 추가된(삭제하지 않은 같은 템플릿 루틴이 있는) 템플릿은
        건너뛴다. 이름은 프로필 언어 값(없으면 한국어), 알림은 꺼짐, 요일 이력은 요청의 현지 날짜부터다. 빈 배열은 건너뛰기이며, 처음 호출이면 selection_completed를
        true로 만든다. 없거나 비활성인 id가 하나라도 있으면 아무것도 만들지 않고 422다.
      parameters:
      - name: X-App-Timezone
        in: header
        required: false
        description: 현재 기기의 IANA 시간대. 생략 시 저장 시간대. 잘못된 값은 APP_TIMEZONE_INVALID.
        schema:
          type: string
          minLength: 1
          maxLength: 64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoutineTemplateSelectionRequest'
      responses:
        '200':
          description: 이번에 만든 루틴
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoutineListResponse'
          headers:
            X-App-Local-Date:
              description: 실제 계산에 사용한 현지 날짜
              schema:
                type: string
                format: date
            X-App-Served-At:
              description: 단일 요청 서버 UTC 시각
              schema:
                type: string
                format: date-time
            X-App-Day-Ends-At:
              description: 다음 현지 달력일 자정의 UTC 시각
              schema:
                type: string
                format: date-time
        '500':
          $ref: '#/components/responses/Internal'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/ProfileNotFound'
        '422':
          description: 요청 형식 오류 또는 없거나 비활성인 템플릿 id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              examples:
                requestValidation:
                  summary: 요청 검증
                  value:
                    error:
                      code: VALIDATION
                      message: 요청 형식이 올바르지 않습니다.
                      details:
                        errors:
                        - loc:
                          - body
                          - template_ids
                          type: value_error
                          msg: Value error, template_ids는 중복 없이 보내야 해요.
                unknownTemplate:
                  summary: 없거나 비활성인 템플릿
                  value:
                    error:
                      code: VALIDATION
                      message: 추가할 수 없는 템플릿이에요.
                      details:
                        template_ids:
                        - unknown_template
...
components:
  schemas:
...
    RoutineHistoryResponse:
      type: object
      properties:
        date:
          type: string
          format: date
        data:
          type: array
          items:
            $ref: '#/components/schemas/RoutineHistoryItem'
      required:
      - date
      - data
      title: RoutineHistoryResponse
    RoutineListResponse:
      properties:
        data:
          items:
            $ref: '#/components/schemas/RoutineResponse'
          type: array
          title: Data
      type: object
      required:
      - data
      title: RoutineListResponse
...
    ErrorEnvelope:
      type: object
      required:
      - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
...
  securitySchemes:
    SupabaseBearer:
      type: http
      scheme: bearer
      bearerFormat: Supabase access token JWT
      description: JWKS로 ES256/RS256 서명, aud=authenticated, sub/exp 필수, leeway=60초를 검증한다. SUPABASE_URL이
        있으면 issuer=<SUPABASE_URL>/auth/v1도 검증한다. is_anonymous=true token은 ALLOW_ANONYMOUS_AUTH 설정(기본 꺼짐)이
        켜진 환경에서만 허용하고 그 외에는 거부한다.
    RevenueCatAuthorization:
      type: apiKey
      in: header
      name: Authorization
      description: RevenueCat dashboard의 전체 Authorization 문자열과 server secret을 상수시간 exact compare한다. Bearer
        prefix 규칙은 없고 설정이 비어도 fail-closed 401이다.
    AdMobSsvSignature:
      type: apiKey
      in: query
      name: signature
      description: URL-safe base64 Google AdMob SSV ECDSA P-256/SHA-256 signature. key_id로 공개키를 선택하고 signature
        앞의 percent-encoded 원본 query를 검증한다. signature와 key_id는 각각 정확히 한 번, query의 마지막 두 필드에 이 순서로 있어야 하며
        그 뒤 suffix는 거절한다.
      x-verification:
        algorithm: ECDSA P-256 SHA-256
        signed_material: signature parameter 앞의 percent-encoded raw query
        public_key_selector: key_id
        envelope: signature와 key_id가 마지막 두 필드이며 suffix와 critical duplicate는 허용하지 않음
  responses:
...
    Unauthorized:
      description: Bearer token 누락, 형식 오류, 검증 실패 또는 허용되지 않은 anonymous token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            unauthorized:
              summary: 인증 실패
              value:
                error:
                  code: UNAUTHORIZED
                  message: 다시 로그인해 주세요.
                  details: {}
    Validation:
      description: 요청 형식 또는 endpoint validation 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            requestValidation:
              summary: 요청 검증
              value:
                error:
                  code: VALIDATION
                  message: 요청 형식이 올바르지 않습니다.
                  details:
                    errors:
                    - loc:
                      - body
                      - field
                      type: missing
                      msg: Field required
    Internal:
      description: DB, 외부 API, 응답 직렬화 등을 포함한 미처리 또는 명시적 내부 오류
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            internal:
              summary: 내부 오류
              value:
                error:
                  code: INTERNAL
                  message: 일시적인 오류가 발생했어요. 잠시 후 다시 시도해 주세요.
                  details: {}
    ProfileNotFound:
      description: 인증 user의 profile이 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            notFound:
              summary: 프로필 없음
              value:
                error:
                  code: NOT_FOUND
                  message: 프로필을 찾을 수 없어요.
                  details: {}
...
servers:
- url: https://voice.moly.asia
  description: 운영 API 서버
- url: http://localhost:8000
  description: 로컬 개발 서버
...